+++ CHANGES.md
@@ -8,6 +8,13 @@ Version 1.6.6
 
 To be released.
 
+ -  Fixed `TypeError: unusable` error that occurred when `doubleKnock()`
+    encountered redirects during HTTP signature retry attempts.
+    [[#294], [#295]]
+
+[#294]: https://github.com/fedify-dev/fedify/issues/294
+[#295]: https://github.com/fedify-dev/fedify/pull/295
+
 
 Version 1.6.5
 -------------